timeseries example - reduce buffersize
make board bufferSize independent from the number of samples drawn on screen.  before that was 1250 by default and that is unnecessary.  much lower cpu cost now.  default set to 512 and that is what the filters will use.  samples on screen still 1250 by default (5sec*250sr).
also fixed+simplified uVrms and now an empty line is drawn when channel inactive.